requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via aiogram
aiogram==3.22.0
    # via auto-steam-gifts (pyproject.toml)
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via aiogram
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.13.4
    # via auto-steam-gifts (pyproject.toml)
certifi==2025.8.3
    # via
    #   aiogram
    #   requests
charset-normalizer==3.4.3
    # via requests
emoji==2.14.1
    # via auto-steam-gifts (pyproject.toml)
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via aiogram
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via auto-steam-gifts (pyproject.toml)
requests==2.32.4
    # via steamspypi
soupsieve==2.7
    # via beautifulsoup4
steamspypi==1.1.1
    # via auto-steam-gifts (pyproject.toml)
tenacity==9.1.2
    # via auto-steam-gifts (pyproject.toml)
typing-extensions==4.14.1
    # via
    #   aiogram
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
yarl==1.20.1
    # via aiohttp
